requirements.txt
langchain>=0.1.0
langchain-pinecone>=0.1.0
langchain-google-genai>=0.1.0
pinecone
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
pypdf>=3.0.0
google-generativeai>=0.3.0
flask>=2.3.0
gunicorn>=21.0.0